Fix backend: riorganizzazione entity in model/ + risoluzione 6 test rotti

+++ db/init.sql
@@ -1,15 +1,9 @@
 -- ArcheType-LBP Database Schema
--- Based on ER diagram (db 1.jpeg), Logical Modeling (db 2.jpeg), Pivot tables (db 4.jpeg)
 
--- Enable UUID extension
 CREATE EXTENSION IF NOT EXISTS "pgcrypto";
 
--- ============================================================
--- CORE TABLES
--- ============================================================
-
 CREATE TABLE IF NOT EXISTS users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     username VARCHAR(50) UNIQUE NOT NULL,
     email VARCHAR(100) UNIQUE NOT NULL,
     password VARCHAR(255) NOT NULL,
@@ -21,37 +15,37 @@ CREATE TABLE IF NOT EXISTS users (
 );
 
 CREATE TABLE IF NOT EXISTS developers (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) UNIQUE NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE IF NOT EXISTS publishers (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255) UNIQUE NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE IF NOT EXISTS genres (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(100) UNIQUE NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE IF NOT EXISTS categories (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(100) UNIQUE NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE IF NOT EXISTS games (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     appid INTEGER UNIQUE NOT NULL,
     name VARCHAR(255) NOT NULL,
     price DECIMAL(10,2) DEFAULT 0.00,
     release_date DATE,
-    developer_id INTEGER REFERENCES developers(id) ON DELETE SET NULL,
-    publisher_id INTEGER REFERENCES publishers(id) ON DELETE SET NULL,
+    developer_id BIGINT REFERENCES developers(id) ON DELETE SET NULL,
+    publisher_id BIGINT REFERENCES publishers(id) ON DELETE SET NULL,
     rating DECIMAL(3,2),
     description TEXT,
     header_image_url VARCHAR(500),
@@ -60,37 +54,28 @@ CREATE TABLE IF NOT EXISTS games (
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
--- ============================================================
--- PIVOT TABLES (Many-to-Many)
--- ============================================================
-
 CREATE TABLE IF NOT EXISTS game_genres (
-    game_id INTEGER REFERENCES games(id) ON DELETE CASCADE,
-    genre_id INTEGER REFERENCES genres(id) ON DELETE CASCADE,
+    game_id BIGINT REFERENCES games(id) ON DELETE CASCADE,
+    genre_id BIGINT REFERENCES genres(id) ON DELETE CASCADE,
     PRIMARY KEY (game_id, genre_id)
 );
 
 CREATE TABLE IF NOT EXISTS game_categories (
-    game_id INTEGER REFERENCES games(id) ON DELETE CASCADE,
-    category_id INTEGER REFERENCES categories(id) ON DELETE CASCADE,
+    game_id BIGINT REFERENCES games(id) ON DELETE CASCADE,
+    category_id BIGINT REFERENCES categories(id) ON DELETE CASCADE,
     PRIMARY KEY (game_id, category_id)
 );
 
--- Conditional: only if a game can have multiple developers
 CREATE TABLE IF NOT EXISTS game_developers (
-    game_id INTEGER REFERENCES games(id) ON DELETE CASCADE,
-    developer_id INTEGER REFERENCES developers(id) ON DELETE CASCADE,
+    game_id BIGINT REFERENCES games(id) ON DELETE CASCADE,
+    developer_id BIGINT REFERENCES developers(id) ON DELETE CASCADE,
     PRIMARY KEY (game_id, developer_id)
 );
 
--- ============================================================
--- USER-GAME RELATIONSHIPS
--- ============================================================
-
 CREATE TABLE IF NOT EXISTS wishlist (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
-    game_id INTEGER REFERENCES games(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
+    game_id BIGINT REFERENCES games(id) ON DELETE CASCADE,
     priority INTEGER DEFAULT 0 CHECK (priority IN (0, 1, 2)),
     notes TEXT,
     added_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -98,9 +83,9 @@ CREATE TABLE IF NOT EXISTS wishlist (
 );
 
 CREATE TABLE IF NOT EXISTS backlog (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
-    game_id INTEGER REFERENCES games(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
+    game_id BIGINT REFERENCES games(id) ON DELETE CASCADE,
     status VARCHAR(20) NOT NULL DEFAULT 'wishlist'
         CHECK (status IN ('wishlist', 'playing', 'finished', 'abandoned')),
     play_time_min INTEGER DEFAULT 0,
@@ -111,34 +96,26 @@ CREATE TABLE IF NOT EXISTS backlog (
 );
 
 CREATE TABLE IF NOT EXISTS game_sessions (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
-    game_id INTEGER REFERENCES games(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
+    game_id BIGINT REFERENCES games(id) ON DELETE CASCADE,
     session_start TIMESTAMP NOT NULL,
     session_end TIMESTAMP,
     duration_min INTEGER,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
--- ============================================================
--- SOCIAL
--- ============================================================
-
 CREATE TABLE IF NOT EXISTS friends (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
-    friend_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
+    friend_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
     status VARCHAR(20) DEFAULT 'pending' CHECK (status IN ('pending', 'accepted', 'blocked')),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     UNIQUE(user_id, friend_id)
 );
 
--- ============================================================
--- USER SETTINGS
--- ============================================================
-
 CREATE TABLE IF NOT EXISTS user_settings (
-    user_id INTEGER PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
+    user_id BIGINT PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
     theme VARCHAR(20) DEFAULT 'dark',
     language VARCHAR(5) DEFAULT 'en',
     colorblind_mode BOOLEAN DEFAULT FALSE,
@@ -146,60 +123,43 @@ CREATE TABLE IF NOT EXISTS user_settings (
     show_wishlist BOOLEAN DEFAULT TRUE
 );
 
--- ============================================================
--- INDEXES
--- ============================================================
-
 CREATE INDEX idx_games_appid ON games(appid);
 CREATE INDEX idx_games_name ON games(name);
 CREATE INDEX idx_games_release_date ON games(release_date);
 CREATE INDEX idx_games_rating ON games(rating);
 CREATE INDEX idx_games_developer ON games(developer_id);
 CREATE INDEX idx_games_publisher ON games(publisher_id);
-
 CREATE INDEX idx_game_genres_game ON game_genres(game_id);
 CREATE INDEX idx_game_genres_genre ON game_genres(genre_id);
-
 CREATE INDEX idx_game_categories_game ON game_categories(game_id);
 CREATE INDEX idx_game_categories_cat ON game_categories(category_id);
-
 CREATE INDEX idx_wishlist_user ON wishlist(user_id);
 CREATE INDEX idx_backlog_user ON backlog(user_id);
 CREATE INDEX idx_backlog_user_status ON backlog(user_id, status);
 CREATE INDEX idx_game_sessions_user ON game_sessions(user_id);
 CREATE INDEX idx_friends_user ON friends(user_id);
 
--- ============================================================
--- TEST DATA (from db 3.jpeg - Verificare con dati di test)
--- ============================================================
-
--- 2 users
 INSERT INTO users (username, email, password) VALUES
     ('gamer_alice', 'alice@example.com', 'password123'),
     ('gamer_bob', 'bob@example.com', 'password123')
 ON CONFLICT (username) DO NOTHING;
 
--- Developers
 INSERT INTO developers (name) VALUES
     ('Valve'), ('CD Projekt Red'), ('FromSoftware'), ('Rockstar Games')
 ON CONFLICT (name) DO NOTHING;
 
--- Publishers
 INSERT INTO publishers (name) VALUES
     ('Valve'), ('CD Projekt'), ('Bandai Namco'), ('Rockstar Games')
 ON CONFLICT (name) DO NOTHING;
 
--- 3 genres
 INSERT INTO genres (name) VALUES
     ('Action'), ('RPG'), ('Strategy')
 ON CONFLICT (name) DO NOTHING;
 
--- Categories
 INSERT INTO categories (name) VALUES
     ('Singleplayer'), ('Multiplayer'), ('Co-op'), ('Open World')
 ON CONFLICT (name) DO NOTHING;
 
--- 5 games
 INSERT INTO games (appid, name, price, release_date, developer_id, publisher_id, rating, description) VALUES
     (730, 'Counter-Strike 2', 0.00, '2012-08-21', 1, 1, 4.50, 'Counter-Strike 2 is the next chapter of the world''s favorite competitive FPS.'),
     (1086940, 'Baldur''s Gate 3', 59.99, '2023-08-03', 2, 2, 4.80, 'An epic RPG from the creators of Divinity: Original Sin 2.'),
@@ -208,29 +168,24 @@ INSERT INTO games (appid, name, price, release_date, developer_id, publisher_id,
     (570, 'Dota 2', 0.00, '2013-07-09', 1, 1, 4.20, 'Every day, millions of players worldwide enter battle as one of over a hundred Dota heroes.')
 ON CONFLICT (appid) DO NOTHING;
 
--- Game-Genre relationships
 INSERT INTO game_genres (game_id, genre_id) VALUES
     (1, 1), (2, 2), (2, 1), (3, 2), (3, 1), (4, 1), (4, 3), (5, 3)
 ON CONFLICT DO NOTHING;
 
--- Game-Category relationships
 INSERT INTO game_categories (game_id, category_id) VALUES
     (1, 2), (1, 3), (2, 1), (2, 3), (3, 1), (3, 4), (4, 1), (4, 2), (4, 4), (5, 2), (5, 3)
 ON CONFLICT DO NOTHING;
 
--- Wishlist for user 1
 INSERT INTO wishlist (user_id, game_id, priority) VALUES
     (1, 2, 2), (1, 3, 1)
 ON CONFLICT DO NOTHING;
 
--- Backlog for user 1
 INSERT INTO backlog (user_id, game_id, status, play_time_min) VALUES
     (1, 1, 'playing', 1200),
     (1, 4, 'finished', 3600),
     (1, 5, 'abandoned', 45)
 ON CONFLICT DO NOTHING;
 
--- Backlog for user 2
 INSERT INTO backlog (user_id, game_id, status, play_time_min) VALUES
     (2, 1, 'finished', 2400),
     (2, 2, 'playing', 800)
